fix(acoustics): self-contained fpcalc bundle + AcoustID bug fixes
- Rewrote Scripts/build-fpcalc.sh to recursively copy all Homebrew
  dylib deps (libavcodec, libavformat, libavutil, libswresample, and
  transitive codec libs) into Resources/, rewrite every Homebrew-
  absolute path to @loader_path, and re-sign; works on macOS bash 3.2.
- Added 'make bundle-fpcalc' target; wired into 'make bootstrap' so
  new devs get self-contained binaries after a single command.
- Added 'chromaprint' to Brewfile; CI runs 'make bundle-fpcalc' before
  'make generate' so XcodeGen sees all dylibs when scanning Resources/.
- Gitignored Resources/fpcalc and Resources/*.dylib (generated artefacts);
  removed the previously-committed fpcalc and libchromaprint.1.dylib.
- Added com.apple.security.network.client entitlement so sandboxed app
  can reach api.acoustid.org.
- Fixed FpcalcOutput.duration and AcoustIDRecording.duration from Int
  to Double (fpcalc and AcoustID API both emit fractional seconds).
- Title-case genre tags from MusicBrainz (MB tags are lowercase).
- Fixed swiftlint trailing_closure violation in TagEditorViewModel.
- Updated README and DEVELOPMENT.md with fpcalc setup, CI, and
  distribution-signing instructions.

+++ Modules/Acoustics/Sources/Acoustics/FingerprintResult.swift
@@ -61,7 +61,7 @@ public struct AcoustIDResult: Decodable, Sendable {
 public struct AcoustIDRecording: Decodable, Sendable {
     public let id: String
     public let title: String?
-    public let duration: Int?
+    public let duration: Double?
     public let artists: [AcoustIDArtist]?
     public let releases: [AcoustIDRelease]?
 }
@@ -105,5 +105,5 @@ public struct AcoustIDTrackPosition: Decodable, Sendable {
 
 struct FpcalcOutput: Decodable {
     let fingerprint: String
-    let duration: Int
+    let duration: Double
 }
